feat: add opt-in Loki query cost guardrail for query_loki_logs
Loki query cost is bytes scanned, determined only by the stream
selector and time range: line filters and parsers reduce what is
returned, not what is read, and Loki does not enforce
max_query_bytes_read on log queries without a line filter. LLM callers
routinely issue catch-all selectors over wide windows, which can scan
terabytes per call.

Gate query_loki_logs behind a three-mode guardrail (off by default;
shadow logs violations; enforce blocks): require a selective positive
label matcher in every stream selector, cap the effective time range
including range-vector durations, and pre-check Loki's index/stats
byte estimate against a budget, counting repeated selectors once per
occurrence with saturating arithmetic. Blocked queries return an MCP
tool error with rewrite guidance so models can retry cheaply;
unparseable queries fail open (Loki reports its own syntax errors) and
are logged. Scanning strips LogQL # comments and handles fractional
range-vector durations so commented-out or fractional syntax cannot
produce false blocks or bypasses. On VictoriaLogs the static checks
apply to selector-shaped queries and the byte pre-check is skipped,
since its stats approximation runs a real query.

Ref grafana/mcp-grafana#1030

+++ tools/loki_test.go
@@ -54,7 +54,7 @@ func TestLokiTools(t *testing.T) {
 		assert.GreaterOrEqual(t, result.Streams, 0, "Should have a valid streams count")
 		assert.GreaterOrEqual(t, result.Chunks, 0, "Should have a valid chunks count")
 		assert.GreaterOrEqual(t, result.Entries, 0, "Should have a valid entries count")
-		assert.GreaterOrEqual(t, result.Bytes, 0, "Should have a valid bytes count")
+		assert.GreaterOrEqual(t, result.Bytes, int64(0), "Should have a valid bytes count")
 	})
 
 	t.Run("query loki logs", func(t *testing.T) {
